requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
sqlparse==0.5.3
